+++ components/Table/TableIndex/TableIndex.tsx
@@ -1,17 +1,17 @@
-import { InviteBoard, Member, Tableindex } from "@/components/Table/Table.type";
-import InviteButton from "@/components/Table/TablePagenation/InviteButton";
+import { InviteBoard, Member, TableIndex } from "@/components/Table/Table.type";
+import InviteButton from "@/components/Table/TablePagination/InviteButton";
 import clsx from "clsx";
 import styles from "./TableIndex.module.css";
 
 interface TableIndexProps {
   data: (Member | InviteBoard)[];
-  tableindex: Tableindex;
+  tableIndex: TableIndex;
   invite?: boolean;
 }
 
-const TableIndex = ({ tableindex, invite = false }: TableIndexProps) => {
-  const column = Object.keys(tableindex).length;
-  const isAccept = Object.values(tableindex).includes("acceptButton");
+const TableIndex = ({ tableIndex, invite = false }: TableIndexProps) => {
+  const column = Object.keys(tableIndex).length;
+  const isAccept = Object.values(tableIndex).includes("acceptButton");
 
   return (
     <div
@@ -20,13 +20,13 @@ const TableIndex = ({ tableindex, invite = false }: TableIndexProps) => {
         gridTemplateColumns: `repeat(${column - 1}, minmax(max-content, 1fr)) ${invite ? "auto" : `minmax(1rem,8rem)`}`,
       }}
     >
-      {Object.keys(tableindex).map(
+      {Object.keys(tableIndex).map(
         (index) =>
           index && (
             <p
               className={clsx(
                 styles.tableindex__item,
-                tableindex[index] === "acceptButton" && styles.tableindex__acceptbutton
+                tableIndex[index] === "acceptButton" && styles.tableindex__acceptbutton
               )}
               key={index}
             >
